removed schema changes, moved components
