+++ packages/fast-component-explorer/app/components/group/index.ts
@@ -1,11 +1,4 @@
-import { TestComponentViewConfig } from "../data.props";
 import groupSchema from "./group.schema";
 import { Group } from "./group";
 
-const group: TestComponentViewConfig = {
-    schema: groupSchema,
-    component: Group,
-};
-
-export { group, groupSchema };
-export * from "./group";
+export { Group, groupSchema };
