+++ packages/fast-component-explorer/app/config.ts
@@ -1,7 +1,7 @@
 import { MenuItem } from "@microsoft/fast-tooling-react";
 import { FormChildOptionItem } from "@microsoft/fast-tooling-react/dist/form/form";
 import { pascalCase } from "@microsoft/fast-web-utilities";
-import * as testComponentViewConfigs from "./utilities/components";
+import * as testComponentViewConfigs from "./components";
 import * as componentViewConfigs from "./utilities/configs";
 import { createBrowserHistory } from "history";
