+++ packages/fast-component-explorer/app/utilities/configs/action-trigger.ts
@@ -5,7 +5,7 @@ import {
     ActionTriggerProps,
     actionTriggerSchema,
 } from "@microsoft/fast-components-react-msft";
-import { glyphSchema, Icon } from "../../../app/components/";
+import { glyphSchema, Icon } from "../../../app/components/glyph";
 import Guidance from "../../.tmp/action-trigger/guidance";
 
 const actionTriggerConfig: ComponentViewConfig<ActionTriggerProps> = {
