+++ packages/fast-component-explorer/app/utilities/configs/text-action.ts
@@ -6,7 +6,7 @@ import {
     TextActionProps,
     textActionSchema,
 } from "@microsoft/fast-components-react-msft";
-import { glyphSchema, Icon } from "../../../app/components/";
+import { glyphSchema, Icon } from "../../../app/components/glyph";
 import Guidance from "../../.tmp/text-action/guidance";
 
 const textActionConfig: ComponentViewConfig<TextActionProps> = {
